# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReviewPlate API - Smart NFC/QR review card platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # USERS
  /users/me:
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # BUSINESS PROFILES
...
  /scan-logs:
    get:
      operationId: listScanLogs
      tags: [scan-logs]
      summary: List scan logs for current user's cards
      parameters:
        - name: cardId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of scan logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanLog"

  # ANALYTICS / DASHBOARD SUMMARY (wow endpoints)
...
  /analytics/cards/{id}/stats:
    get:
      operationId: getCardStats
      tags: [scan-logs]
      summary: Stats for a specific card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Card stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardStats"

  # AI REPLY
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        plan:
          type: string
        language:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - plan
        - language
        - createdAt
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - name
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required:
        - token
        - password
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
...
    ScanLog:
      type: object
      properties:
        id:
          type: integer
        cardId:
          type: integer
        cardCode:
          type: string
        timestamp:
          type: string
        country:
          type: ["string", "null"]
        deviceType:
          type: ["string", "null"]
        ratingGiven:
          type: ["number", "null"]
        wasNegative:
          type: boolean
      required:
        - id
        - cardId
        - cardCode
        - timestamp
        - wasNegative
...
    CardStats:
      type: object
      properties:
        cardId:
          type: integer
        totalScans:
          type: integer
        scansThisMonth:
          type: integer
        scansThisWeek:
          type: integer
        averageRating:
          type: ["number", "null"]
        negativeCount:
          type: integer
        scansByDay:
          type: array
          items:
            $ref: "#/components/schemas/DayStat"
        scansByCountry:
          type: array
          items:
            $ref: "#/components/schemas/CountryStat"
      required:
        - cardId
        - totalScans
        - scansThisMonth
        - scansThisWeek
        - negativeCount
        - scansByDay
        - scansByCountry
    DayStat:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count
    CountryStat:
      type: object
      properties:
        country:
          type: string
        count:
          type: integer
      required:
        - country
        - count
...